+++ src/index.js
@@ -1,9 +1,12 @@
 const VRT = require('./lib/vrt')
 const OGR = require('./lib/ogr')
+const random = require('randomstring')
+const rimraf = require('rimraf')
 const _ = require('highland')
 
 function createStream (format, options) {
-  const stream =  _.pipeline(stream => {
+  options.path = `${options.path || '.'}/${random.generate()}`
+  const stream = _.pipeline(stream => {
     return stream
     .pipe(VRT.createStream(options))
     .on('log', l => stream.emit('log', l))
@@ -11,6 +14,9 @@ function createStream (format, options) {
     .pipe(OGR.createStream(format, options))
     .on('log', l => stream.emit('log', l))
     .on('error', e => stream.emit('error', e))
+    .on('end', e => rimraf(options.path, err => {
+      if (err) stream.emit('log', {level: 'error', message: 'Failed to delete temporary directory'})
+    }))
   })
   return stream
 }